Repair int32 tests.

+++ SingleSource/Regression/C/gcc-c-torture/execute/20071211-1.c
@@ -1,3 +1,5 @@
+// Modified by LLVM-MOS.
+
 extern void abort() __attribute__ ((noreturn));
 
 struct s
@@ -12,8 +14,8 @@ struct s
 
 int main()
 {
-  int f2;
-  sv.f2 = (1 << 24) - 1;
+  long f2;
+  sv.f2 = (1L << 24) - 1;
   __asm__ volatile ("" : : : "memory");
   ++sv.f2;
   f2 = sv.f2;
